README
Program Description:

There are two main components in this project. One is the client side and the other is server side. 

•	Client Side:
           ENC is a client interface made using HTML, CSS and JavaScript for creating text areas with input fields. All the encryptions and decryptions are done at the client side. For these encryptions and decryptions I am using AES256 key encryption algorithm which is a        symmetric key encryption algorithm i.e. same key is used for both encryption and decryption.

•	Server Side:
           CANA is a major component at the server side which is a Restful API written in a Google “GO” language. Ideally CANA should run on dedicated portable hardware like a raspberry pi, for instance it is possible to simply access via bluetooth and a modern browser once configured btnap, but precompiled versions are available for several operating systems and ENChitectures. Therefore, enc can run on pretty much everything with a CPU, from your smartphone, your computer. As a rule of thumb, the more isolated the hardwares, the better.

The client interface allows the users to encrypt the files by uploading it into the raspberry pi with a specific key i.e. password and stores it in a sqlite database. Whenever a user is willing to access the files, the same key should be provided in order to decrypt the files stored in a database via CANA which acts as mediator between client interface and database.

Setup Directions:

In this project as a hardware device,  will be using raspberry pi. Below are the steps to be followed to setup raspberry pi
•	Format a micro sd card and install Raspbian on it ( Download ISO, Verify, dd, Mount).
•	Enable ethernet connectivity over its USB port.
•	Unmount the micro sd card and insert it into the RPIZ and using the data port plug into the computer.
•	Now the computer should detect the network interface.
•	It should be assigned a static IP address in order to connect to it. 
•	Restart the interface and the RPIZ should be reachable. In order to connect it via SSH  used command “ssh pi@raspberrypi.local” . The board has been setup and ready to use.
•	Now install the google “Go” Language on raspberry pi and make sure it is configured correctly.

Usage:
ENC is a client interface implemented in javascript and Html5 that uses CANA, a RESTful API server i.e. written Go language which provides permissions to Read and Write for encrypted records in sqlite database file. 
Records are generated, encrypted and decrypted client side only i.e ENC which relies on CryptoJS for the sake of AES encryption and Pseudo Random Number Generator. ENC, a client side application offers the following UI widgets:
•	Simple Text areas
•	A password field with password strength estimation and a random password generator.
•	Simple Text Inputs
•	Custom File Attachments where the files are encrypted at client side before they are uploaded as binary records
•	Make sure that all the files of ENC, CANA and configuration files are on the RPIZ. Open the browser and go to http://raspberrypi.local.8080/, and should now be able to login with the credentials provided in the config.json file ( contains details about address, port, username, password, database, encryption key which is used to connect to the ENC web interface) and use client side interface whenever raspberry pi is connected to USB port.
Security Considerations:
•	As all the data is encrypted on the client side and store physically on RPIZ which uses AES encryption, we should use a stronger encryption key. The stronger the key, The safer the data.
•	Don’t enable any kind of connection sharing from computer to RPIZ as we don’t want any other illegal sources to reach our secure storage.
•	Username and password are needed to access the API itself, but they will not decrypt the records, that’s why the encryption key is requested as well. You can login with the same API credentials but different encryption keys, you will create records with a new key and will not be able to decrypt other records that have been created with a different AES key.

Major Components:

The major components here in our project are the Client and the server. As mentioned above for the client side I am using an interface called ENC which is the extended idea which is an open source from the “Advanced Rest Client Developers” (https://github.com/advanced-rest-client/ENC-dev-guide). For the server side using a RESTful API called CANA which I referenced from a website “https://dev.to/codehakase/building-a-restful-api-with-go”. The main idea for the password manager is referenced from “GoPass” which is a password manager for teams build using ‘GO’ and ‘UNIX’ operating system (“https://github.com/gopasspw/gopass”).